Lots of codegen made. Nothing tested yet

+++ j--/tests/pass/Division.java
@@ -2,10 +2,9 @@ package pass;
 import java.lang.System;
 
 public class Division {
-    public int divide(int x, int y) {
+    public double divide(double x, double y) {
         return x / y;
     }
-
     //public static void main (String[] args) {
     //    System.out.println(divide(10, 2));
     //}
